Added doc and getters
